builtins(#5714): random_bytes returns u8[] instead of string
random_bytes(n) was the last byte-producing stdlib surface still using
string as a raw byte bag (its result is essentially never valid UTF-8),
blocking D9's validity invariant (#5634, epic #5626). It now returns a
fresh u8[] in the __alloc_u8 box shape on every backend:

- checker: signature is (n: number) -> u8[]
- interp: Array of Number bytes
- x86-64 / arm64: __fern_random_bytes allocs via __alloc_u8 (cap/rc/len
  header, n==0 empty sentinel) then getrandom / chunked getentropy; the
  arm64 two-word-string and trailing-NUL paths are gone
- arm64ssa: inline bump-alloc of the 16-byte-header array box
- wasm: helper returns a single-word array pointer on both preview-1 and
  preview-2 (p2 allocs padded capacity for its u64 stores, then corrects
  the length prefix)
- self-host: rt_src_random_bytes fills a scratch via __syscall3 and
  spreads the bytes into a __raw_arr_box; irlower moves random_bytes
  from string-tracking to array-tracking (both hand-maintained lists,
  per the #5983 split-brain warning) and untaints it like __alloc_u8;
  the self-host wasm IR helper builds the [len][cap][elems] array block
  the preview-2 body already used

Because the result now carries a real rc header on every backend, the
Perceus special cases guarding the old raw unheadered buffer are deleted
(rc_analysis "must stay ineligible" arm, stale rc_insert/ir.go comment)
and the native result is reclaimed: the x86-64 sanitizer shows
allocs=200 frees=200 for a 200-draw loop, and the self-host goes from
freeing none of 400 blocks to freeing the 200 array boxes (the rt_src
scratch leak is pre-existing and shared with read_file). A short
getrandom read now yields zero bytes instead of uninitialised heap.

Callers migrate mechanically (std/uuid, std/math, ~30 test programs);
b[i] as i32 casts were already in place almost everywhere. write()/
write_file() consumers bridge via string_from_bytes_unchecked. The
i32[] annotations that only compiled because the self-host front end
does not type builtin returns are retyped u8[].

Gates run: checker/interp/ir/caps/platforms units; wasmbin emit + sig
parity; e2e Random/LenCallReceiver + rc corpus (x86-64, arm64, wasm);
e2eselfhost math/uuid/component-random/rc-wasm/audit/runtime-helper and
the x86-64+wasm+arm64 IR path suites; feature census; sourcelint;
stdlib check; 7-lane manual differential of the uuid + chunked-600-byte
shapes.

tcp_recv and read_file remain; #5714 stays open for them.

+++ internal/e2eselfhost/self_host_wasm_component_ir_test.go
@@ -130,7 +130,7 @@ func TestSelfHostWasmComponentIRPath(t *testing.T) {
 		// positionally.
 		{"io-random-i32", true, `function main(): i32 { if (random_i32() != 0) { write("r"); } return 0; }`, true,
 			[]string{"wasi:cli/stdout@0.2.0 get-stdout", "wasi:io/streams@0.2.0 [method]output-stream.blocking-write-and-flush", "wasi:random/random@0.2.0 get-random-u64"}},
-		{"io-random-bytes", true, `function main(): i32 { var b: i32[] = random_bytes(4); if (b.len() == 4) { write("b"); } return 0; }`, true,
+		{"io-random-bytes", true, `function main(): i32 { var b: u8[] = random_bytes(4); if (b.len() == 4) { write("b"); } return 0; }`, true,
 			[]string{"wasi:cli/stdout@0.2.0 get-stdout", "wasi:io/streams@0.2.0 [method]output-stream.blocking-write-and-flush", "wasi:random/random@0.2.0 get-random-u64"}},
 		{"io-clock-wall", true, `function main(): i32 { if (now_unix_ms() > 0) { write("w"); } return 0; }`, true,
 			[]string{"wasi:cli/stdout@0.2.0 get-stdout", "wasi:io/streams@0.2.0 [method]output-stream.blocking-write-and-flush", "wasi:clocks/wall-clock@0.2.0 now"}},
